Move blog to a new Django app
