+++ server/blog/translation.py
@@ -1,0 +1,14 @@
+from modeltranslation.translator import register, TranslationOptions
+from blog.models import Blog, Tags
+
+
+@register(Tags)
+class TagsTranslation(TranslationOptions):
+    fields = ('name',)
+    required_languages = ('he',)
+
+
+@register(Blog)
+class BlogTranslation(TranslationOptions):
+    fields = ('title', 'author', 'body')
+    required_languages = ('he',)
